rasterToPoints: dealing with nan improved
It has been improved to be faster (now the nas are checked in the
numpy array step) and to deal with more than one column. IF one of the
interesting columns has not a NAN, i.e. has a right value, all the
columns are returned (with the nan if they exist)